Factored out an attribute syntax provider utility
